+++ lib/eggnstone_flutter.dart
@@ -5,12 +5,11 @@ export 'package:eggnstone_flutter/services/analytics/AnalyticsNavigatorObserver.
 export 'package:eggnstone_flutter/services/analytics/IAnalyticsService.dart';
 export 'package:eggnstone_flutter/services/crashlytics/CrashlyticsMixin.dart';
 export 'package:eggnstone_flutter/services/crashlytics/ICrashlyticsService.dart';
-export 'package:eggnstone_flutter/services/logger/LoggerMixin.dart';
-export 'package:eggnstone_flutter/services/logger/LoggerService.dart';
 export 'package:eggnstone_flutter/services/shared_preferences/SharedPreferencesMixin.dart';
 export 'package:eggnstone_flutter/services/shared_preferences/SharedPreferencesService.dart';
 export 'package:eggnstone_flutter/services/url_launcher/UrlLauncherMixin.dart';
 export 'package:eggnstone_flutter/services/url_launcher/UrlLauncherService.dart';
 export 'package:eggnstone_flutter/tools/DartTools.dart';
+export 'package:eggnstone_flutter/tools/LogTools.dart';
 export 'package:eggnstone_flutter/tools/PackageInfo.dart';
 export 'package:eggnstone_flutter/widgets/BusyDialog.dart';